Main Menu Layout + Logic
Basic layout of the application initialized too. Menu does not look as good as intended, but it is very hard to make it look nice and responsive. Currently it is semi responsive and I would say acceptable in terms of looks. Major limitation not being able to make images responsive in a reasonable way