fixed username duplication bug
Creating more than one staff member with blank username was not possible because we would pass empty strings as username, which is unique. Now adding blank username defaults to null instead. Defaulting to empty string is beneficial almost everywhere because it prevents having to add defaults for every non-mandatory field in code, but it was not beneficial here because of that miscalculation.